Refactor setup script

+++ setup.sh
@@ -2,33 +2,45 @@
 
 SETUP_DIR=$(dirname "$(readlink -f "$0")")
 LINK_ONLY=false
+INSTALL_PACKAGE=
 
+# Show usage
 _usage() {
   echo "$0 usage:" && grep " .)\\ #" $0 | sed -e 's/\(.\)) #/-\1/g'
   exit 0;
 }
 
+# Create symbolic link for config
+# $1: The source of symlink
+# $2: The destination of symlink
+# $3: When true, create parent directories to the link as needed. (default: true)
+# $4: Force removes the old destination, file or folder, and forces a new link (default: true)
 _link() {
+  local need_create_parent=${3:-true}
+  local is_force=${4:-true}
+
   if [[ -e "$2" || -L "$2" ]]; then
+  # Check destination is existed or is a symlink
     if test "$1" -ef "$(readlink -f "$2")"; then
+    # If symlink is setup same with source then finish
       echo "${2} linked"
       return
     else
-      rm -rf "$2"
-    fi
-  else
-    if [[ -d $1 && $HOME != "$(dirname "$2")" && ! -d "$(dirname "$2")" ]]; then
-      rm -rf "$(dirname "$2")"
+    # Delete destination if not a symlink or target of symlink is not same with source
+      [ "$is_force" = true ] && rm -rf "$2"
     fi
   fi
-  mkdir -p "$(dirname "$2")"
+  [ "$need_create_parent" = true ] && mkdir -p "$(dirname "$2")"
   ln -sv "$1" "$2"
 }
 
+# Create empty folder
+_create() {
+  mkdir -p "$*"
+}
+
 _install() {
-  if ! $LINK_ONLY; then
-    bash $SETUP_DIR/distros/install_linux_package.sh
-  fi
+  ! $LINK_ONLY && $SETUP_DIR/distros/install_linux_package.sh
 
   # Basic service
   _link "$SETUP_DIR/git" ~/.git
@@ -36,9 +48,7 @@ _install() {
 
   # Become a hacker
   _link "$SETUP_DIR/fish" ~/.config/fish # Must setup fish shell before vim
-  if ! $LINK_ONLY; then
-    $SETUP_DIR/fish/setup.fish
-  fi
+  ! $LINK_ONLY && chsh -s $(which fish) && $SETUP_DIR/fish/setup.fish
   _link "$SETUP_DIR/vim" ~/.config/nvim
   _link "$SETUP_DIR/vim" ~/.vim
   _link "$SETUP_DIR/vim/vimrc" ~/.vimrc
@@ -49,6 +59,7 @@ _install() {
       vim +PlugInstall +qa
     fi
   fi
+  ! $LINK_ONLY && bash $SETUP_DIR/bin/update.sh -n
 
   # Terminal application
   _link "$SETUP_DIR/tmux/config" ~/.tmux.conf
@@ -56,7 +67,7 @@ _install() {
   _link "$SETUP_DIR/bat" ~/.config/bat
 }
 
-while getopts "hls" arg; do
+while getopts "hlsp" arg; do
   case $arg in
     s) # Update git submodule before setup
       git -C $SETUP_DIR submodule update --init --recursive --remote
